// src/types.ts
export interface Token {
  contractId: string;
  metadata: {
    symbol: string;
    name: string;
    decimals: number;
    identifier?: string; // SIP-010 token identifier
    images?: {
      logo?: string;
    };
  };
}

// Interface for SIP-010 compliant LP tokens
export interface LPToken {
  contractId: string;
  metadata: {
    symbol: string;
    name: string;
    decimals: number;
    identifier: string; // Required for LP tokens
  };
}

export interface Pool extends LPToken {
  token0: Token;
  token1: Token;
  poolData: {
    reserve0: number;
    reserve1: number;
    totalSupply: number;
    fee: number;
  };
  metadata: {
    symbol: string;
    name: string;
    decimals: number;
    identifier: string; // LP token identifier is required for pools
    description?: string;
    image?: string;
  };
}